+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- **DSF**: Added support for DSD Stream Format (DSF) files ([PR](https://github.com/Serial-ATA/lofty-rs/pull/718))
+
 ## [0.25.4] - 2026-09-20
 
 ### Added